delete user

+++ app/Http/Controllers/UserController.php
@@ -14,6 +14,9 @@ class UserController extends Controller
     public function delete($id)
     {
         $user = User::findOrFail($id);
+        if(!$user){
+            return response()->json(['message' => 'User not found'], 404);
+        }
         $user->delete();
         return response()->json(['message' => 'User deleted successfully'], 200);
     }
